+++ internal/cli/init.go
@@ -96,6 +96,7 @@ repository:
 
 sops:
   binary: %s
+  age_binary: %s
   age_key_file: %s
 
 git:
@@ -113,7 +114,7 @@ validation:
 profiles:
   %s:
     renders: []
-`, cfg.Version, cfg.Repository.Root, cfg.Repository.EncryptedExtension, cfg.SOPS.Binary, cfg.SOPS.AgeKeyFile, cfg.Git.AutoCommit, cfg.Defaults.OutputFormat, cfg.Defaults.OutputDir, cfg.Defaults.FileMode, cfg.Validation.RequireValues, cfg.Validation.KeyPattern, profile)) + "\n"
+`, cfg.Version, cfg.Repository.Root, cfg.Repository.EncryptedExtension, cfg.SOPS.Binary, cfg.SOPS.AgeBinary, cfg.SOPS.AgeKeyFile, cfg.Git.AutoCommit, cfg.Defaults.OutputFormat, cfg.Defaults.OutputDir, cfg.Defaults.FileMode, cfg.Validation.RequireValues, cfg.Validation.KeyPattern, profile)) + "\n"
 }
 
 func defaultSOPSYAML() string {